move logic to webhook only
